Testando

+++ src/app.ts
@@ -1,10 +1,12 @@
 import express from "express";
-import profileRoutes from "./routes/profile-route.js"; 
-import depositRoutes from "./routes/deposit-route.js"; 
-import jobRoutes from './routes/job-route.js';
-import paymentRoutes from './routes/payment-route.js'
-import contractRoutes from './routes/contract-route.js';
-import sequelize from "./shared/connection.js"; 
+import profileRoutes from "./routes/profile-route"; 
+import depositRoutes from "./routes/deposit-route";
+import jobRoutes from './routes/job-route';
+import paymentRoutes from './routes/payment-route';
+import contractRoutes from './routes/contract-route';
+import sequelize from "./shared/connection"; 
+
+
 
 const app = express();
 app.use(express.json());
@@ -22,22 +24,22 @@ app.use('/api/payment', paymentRoutes);
 
 
 
-(async () => {
-    try {
+if (process.env.NODE_ENV !== 'test') {
+    (async () => {
+        try {
+            await sequelize.authenticate();
+            console.log("Database connected successfully");
 
-    
-        await sequelize.authenticate();
-        console.log("Database connected successfully");
+            await sequelize.sync({ force: false });
+            console.log("Models synchronized with the database.");
 
-        await sequelize.sync({ force: false }); 
-        console.log("Models synchronized with the database.");
-
-        app.listen(PORT, () => {console.log("Server is running on port", PORT);
-            console.log("Server is running on port", PORT);});
-    } catch (error) {
-        console.error("Unable to connect to the database", error);
-    }
-})();
-
+            app.listen(PORT, () => {
+                console.log("Server is running on port", PORT);
+            });
+        } catch (error) {
+            console.error("Unable to connect to the database", error);
+        }
+    })();
+}
 
 export default app;
